// src/models/product.ts
export interface Product {
  productCartId: any;
  productName: string;
  productPrinterName: string;
  productMenuName: string;
  categoryInformation: any;
  variant: any;
  ingredients: any;
  modifiers: any;
  variantTotalAmount: number;
  modifiersTotalAmount: number;
  taxesTotalAmount: number;
  quantity: number;
  notes: string;
  taxes: any;
  subTotalAmount: number;
  totalAmount: number;
}